+++ src/app/models/dto/user.ts
@@ -1,0 +1,6 @@
+export class User {
+    public bsn: number;
+    public loggedIn: boolean;
+    public isInwoner: boolean;
+    public isProfessional: boolean;
+}
