+++ src/app/views/irma/irma.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Renderer2, Input } from '@angular/core';
-import * as jwtDecode from 'jwt-decode';
+// import * as jwtDecode from 'jwt-decode';
 import { HttpHeaders, HttpClient } from '@angular/common/http';
-import { mapTo, delay } from 'rxjs/operators';
-import jwt_decode = require('jwt-decode');
+// import { mapTo, delay } from 'rxjs/operators';
+// import jwt_decode = require('jwt-decode');
 import { AuthService } from '../../services/auth-service';
 
 @Component({
@@ -21,7 +21,7 @@ export class IrmaComponent implements OnInit {
   header: any;
   payload: any;
   signature: any;
-  timer: NodeJS.Timer = null;
+  timer: any;
 
   p1: string;
   p2: string;
